refactor: move v3 files to a dedicated directory
