+++ lib/v3/errorMap.ts
